Rename TabbedContent modules to Tabs
