+++ client/src/components/Tabs/Tabs.scss
