+++ client/src/components/index.js
@@ -9,7 +9,7 @@ export { HeightClipper } from "./HeightClipper/HeightClipper";
 export { RadioButtons } from "./RadioButtons/RadioButtons";
 export { Select } from "./Select/Select";
 export { SortDirections } from "./SortDirection/SortDirection";
-export { Tabs, TabsStateful } from "./TabbedContent/TabbedContent";
+export { Tabs, TabsStateful } from "./Tabs/Tabs";
 export { LabeledBox } from "./LabeledBox/LabeledBox";
 export { TextMaker, TM } from "./TextMaker";
 export { TwoLevelSelect } from "./TwoLevelSelect/TwoLevelSelect";
